Add in-game options how battery banks should be charged
Changes the power distribution logic to take either
from a battery or a generator (which ever is closer).
Sometimes we want diesel only as a backup, which wasn't
possible with previous logic. It was mainly added to
optimize battery bank charging ping-pong. This should
allow users to create better managed power grids.

Since we need to persist these options, the save file
will be incompatible. Added a (dangerous) options to
allow loading vanilla maps to be "upgraded".