Fix EnumGamePrefs to be compatible with any runtime version
Previously we hard-coded the enums during compilation (e.g.
storing the actual int the enum had from Assembly-Csharp.dll
when it was compiled). Later this dll may be used with a new
version that has `EnumGamePrefs` altered, thus the hard-coded
value would not match anymore with the new runtime dll.