[DF] Update for DF 4.1.0 compatibility (for Quartz)
Note that the condition is a bit weird in order to get the
default behavior. This is a bug in the XmlPatcher which I
can't fix without creating a new patcher version. This works
for now until I have the need to make a new version anyway.
Patcher version 3 will have the proper fix for that!